+++ server/src/main/java/com/vip/interviewpartner/common/exception/GlobalExceptionHandler.java
@@ -1,6 +1,6 @@
 package com.vip.interviewpartner.common.exception;
 
-import com.vip.interviewpartner.common.ApiCommonResponse;
+import com.vip.interviewpartner.common.dto.ApiCommonResponse;
 import jakarta.validation.ConstraintViolation;
 import jakarta.validation.ConstraintViolationException;
 import java.util.Set;
